requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
Django==3.1.7
django-allauth==0.44.0
django-channels==0.7.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
Pillow==8.1.2
PyJWT==1.7.1
pytest==6.2.3
requests==2.25.1
selenium==3.141.0
virtualenv==20.4.7asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-filter==2.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-filter==2.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
